split router initialisation and dispatch in two files as router was getting quite big
